api/requirements.txt
flask
flask-cors
flask-mysql
flask-jwt-extended
mysql-connector-python
